requirements.txt
streamlit==1.28.0
opencv-python-headless==4.8.1.78
deepface==0.0.79
numpy==1.23.5
Pillow==10.0.0
tensorflow-cpu==2.12.0
keras==2.12.0
protobuf==3.20.3
setuptools==65.5.1
wheel==0.40.0
python-multipart==0.0.6
gunicorn==20.1.0
Flask==2.0.1
Flask-Cors==3.0.10
mediapipe==0.10.7
dlib==19.24.1
retina-face==0.0.13
mtcnn==0.1.1
opencv-contrib-python-headless==4.8.1.78
scikit-learn==1.3.0
imutils==0.5.4
tqdm==4.66.1
pandas==2.0.3
gdown==4.7.1